Refactor Makefile to remove unnecessary build dependency
Previously, the 'image' target in the Makefile had a dependency on the
'build' target. This dependency was redundant and caused the 'image'
target to execute unnecessary build steps. The updated Makefile removes
this dependency, optimizing the build process for creating Docker
images. This change reduces build time and improves overall efficiency.

Issue: #1234